requirements.txt
google-generativeai
pymupdf
python-dotenv
typer[all]
